Rename `src` to `path` (#184)
Rename `src` to `path` for `NamelistFile`, `ShellTask`, `AvailableData` and
`GeneratedData`. 

We added path checks to the yaml parsing if they are required to be relative or
absolute.

We also include `IconTask.bin` in the representation as this has been made
absolute in PR #153 (commit 8b5ec4da) and thus does not create a path
that is dependent on the local development environment.