Remove `filterwarnings` conflicting with aiida EBM (#201)
The pytest mechanism erroring out for thrown warnings can interrupt aiida in a
transport task. In this case the EBM (exponential backoff mechanism) of aiida
catches the pytest error and blocks the test suite in erroring out. In the new
aiida release transport throws a deprecation waring. This results in
intransparent error state as pytest's error message is never printed out. We
remove from `filterwarning` the error on any warning and will only error out on
certain warnings.